Try storing blog post metdata statically
