+++ app/thoughts/utils.ts
@@ -1,86 +1,15 @@
-import fs from 'fs'
-import path from 'path'
+import posts from '../data/blog-posts.json'
 
-const META = /export\s+const\s+meta\s+=\s+(\{(\n|.)*?\n\})/
-
-type Metadata = {
-  title: string
-  publishedAt: string
-  summary: string
-  state: 'draft' | 'published' | 'archived'
-  views: number
-  image?: string
-}
-
-function parseFrontmatter(fileContent: string) {
-  const match = META.exec(fileContent)
-  if (!match || typeof match[1] !== 'string')
-    throw new Error(`${name} needs to export const meta = {}`)
-
-  const meta = eval('(' + match[1] + ')')
-  let content = fileContent.replace(META, '').trim()
-
-  return {
-    metadata: {
-      // Default views to 0
-      views: 0,
-      ...meta,
-    } as Metadata,
-    content,
-  }
-}
 export function getBlogPosts({
   drafts = false,
   archived = false,
 }: { drafts?: boolean; archived?: boolean } = {}) {
-  let files: string[] = []
+  return posts.filter((post) => {
+    if (drafts && post.metadata.state === 'draft') return true
+    if (archived && post.metadata.state === 'archived') return true
 
-  const subfolders = fs
-    .readdirSync(path.resolve(process.cwd(), './app', 'thoughts'), {
-      withFileTypes: true,
-    })
-    .filter((dirent) => dirent.isDirectory())
-    .map((dirent) => dirent.name)
-
-  for (const subfolder of subfolders) {
-    const subfolderPath = path.join(
-      path.resolve(process.cwd(), './app', 'thoughts'),
-      subfolder
-    )
-    const subfolderFiles = fs
-      .readdirSync(subfolderPath)
-      .filter((file) => file === 'page.mdx')
-      .map((file) => path.join(subfolder, file))
-    files.push(...subfolderFiles)
-  }
-
-  return files
-    .map((file) => {
-      let rawContent = fs.readFileSync(
-        path.join(path.resolve(process.cwd(), './app', 'thoughts'), file),
-        'utf-8'
-      )
-      let { metadata, content } = parseFrontmatter(rawContent)
-      let slug = path.basename(path.dirname(file))
-
-      return {
-        metadata,
-        slug,
-        content,
-      }
-    })
-    .sort((a, b) => {
-      if (new Date(a.metadata.publishedAt) > new Date(b.metadata.publishedAt)) {
-        return -1
-      }
-      return 1
-    })
-    .filter((post) => {
-      if (drafts && post.metadata.state === 'draft') return true
-      if (archived && post.metadata.state === 'archived') return true
-
-      return post.metadata.state === 'published'
-    })
+    return post.metadata.state === 'published'
+  })
 }
 
 export function formatDate(date: string, includeRelative = false) {
